[refactor] #추천코스 장원태 [231231] - courselist.js의 유틸함수 분리
